# Add a benchmark for report and optimize it further.