<!--
Request Prow to automatically lint any go code in this PR:

/lint
-->

<!-- Please include the 'why' behind your changes if no issue exists -->
## Proposed Changes

This might be slightly controversial. As per the test change, the semantics of the stats reported via Prometheus slightly change in that if a scale from 0 request was part of the current reporting period, the prometheus metrics will actually contain the discounted metrics for 1 second. I figured that's not too harmful seeing as the performance gains to the report function are quite high and the scrape interval on prometheus endpoints is kinda low usually anyway, so the chance that anybody sees the wrong value is fairly low anyway.

## Benchmark before dropping the prometheus reporting from hot path

```
goos: linux
goarch: amd64
pkg: knative.dev/serving/pkg/activator/handler
BenchmarkConcurrencyReporterReport/revs-1-16         	  712749	      1623 ns/op	     960 B/op	      17 allocs/op
BenchmarkConcurrencyReporterReport/revs-5-16         	  156891	      7566 ns/op	    4576 B/op	      77 allocs/op
BenchmarkConcurrencyReporterReport/revs-10-16        	   79693	     14943 ns/op	    9105 B/op	     152 allocs/op
BenchmarkConcurrencyReporterReport/revs-50-16        	   15901	     76126 ns/op	   46385 B/op	     752 allocs/op
BenchmarkConcurrencyReporterReport/revs-100-16       	    6978	    152694 ns/op	   92706 B/op	    1502 allocs/op
BenchmarkConcurrencyReporterReport/revs-200-16       	    3963	    305022 ns/op	  188425 B/op	    3004 allocs/op
PASS
```

## Benchmark with the reporting dropped from hot path

```
goos: linux
goarch: amd64
pkg: knative.dev/serving/pkg/activator/handler
BenchmarkConcurrencyReporterReport/revs-1-16         	 4048651	       294 ns/op	     144 B/op	       2 allocs/op
BenchmarkConcurrencyReporterReport/revs-5-16         	 1632193	       728 ns/op	     496 B/op	       2 allocs/op
BenchmarkConcurrencyReporterReport/revs-10-16        	  918271	      1251 ns/op	     944 B/op	       2 allocs/op
BenchmarkConcurrencyReporterReport/revs-50-16        	  206355	      5550 ns/op	    4914 B/op	       2 allocs/op
BenchmarkConcurrencyReporterReport/revs-100-16       	  112792	     10861 ns/op	    9528 B/op	       2 allocs/op
BenchmarkConcurrencyReporterReport/revs-200-16       	   57441	     20658 ns/op	   18514 B/op	       2 allocs/op
PASS
```

## Compare

```
benchmark                                          old ns/op     new ns/op     delta
BenchmarkConcurrencyReporterReport/revs-1-16       1623          294           -81.89%
BenchmarkConcurrencyReporterReport/revs-5-16       7566          728           -90.38%
BenchmarkConcurrencyReporterReport/revs-10-16      14943         1251          -91.63%
BenchmarkConcurrencyReporterReport/revs-50-16      76126         5550          -92.71%
BenchmarkConcurrencyReporterReport/revs-100-16     152694        10861         -92.89%
BenchmarkConcurrencyReporterReport/revs-200-16     305022        20658         -93.23%

benchmark                                          old allocs     new allocs     delta
BenchmarkConcurrencyReporterReport/revs-1-16       17             2              -88.24%
BenchmarkConcurrencyReporterReport/revs-5-16       77             2              -97.40%
BenchmarkConcurrencyReporterReport/revs-10-16      152            2              -98.68%
BenchmarkConcurrencyReporterReport/revs-50-16      752            2              -99.73%
BenchmarkConcurrencyReporterReport/revs-100-16     1502           2              -99.87%
BenchmarkConcurrencyReporterReport/revs-200-16     3004           2              -99.93%

benchmark                                          old bytes     new bytes     delta
BenchmarkConcurrencyReporterReport/revs-1-16       960           144           -85.00%
BenchmarkConcurrencyReporterReport/revs-5-16       4576          496           -89.16%
BenchmarkConcurrencyReporterReport/revs-10-16      9105          944           -89.63%
BenchmarkConcurrencyReporterReport/revs-50-16      46385         4914          -89.41%
BenchmarkConcurrencyReporterReport/revs-100-16     92706         9528          -89.72%
BenchmarkConcurrencyReporterReport/revs-200-16     188425        18514         -90.17%
```

And yes: The prometheus stuff is **really** too expensive imo.

**Release Note**

<!-- Enter your extended release note in the below block. If the PR requires
additional action from users switching to the new release, include the string
"action required". If no release note is required, write "NONE". -->

```release-note
NONE
```

/assign @julz @vagababov